Start of the structure for ldmud and mudlib containers
The `build/` directories hold the relevant `Dockerfile`s. Check out the `INSTALL.md` for info on how to build then run them. As-is, this builds and starts two containers, one for the ldmud and one for the mudlib.

That being said, there's a bunch that isn't working because I don't
entirely know how to get ldmud up and running. It seems like there are
some dependencies that are missing that you can probably identify at the
drop of a hat. I figure this could at least start the conversation! To
see the dependency issue, do the following:

1. Edit `build/ldmud/Dockerfile` and uncomment the `autogen.sh && ./configure && make`
1. From the root directory, run: `docker build acmemud/ldmud:3.5.0 build/ldmud`

Prerequisites: You need docker installed and running (of course)